floatie

+++ sky/schemas/api/responses.py
@@ -118,15 +118,16 @@ class StatusResponse(ResponseBaseModel):
     memory: Optional[str] = None
     accelerators: Optional[str] = None
 
+
 class ClusterJobRecord(ResponseBaseModel):
     """Response for the cluster job queue endpoint."""
     job_id: int
     job_name: str
     username: str
     user_hash: str
-    submitted_at: int
-    start_at: int
-    end_at: int
+    submitted_at: float
+    start_at: float
+    end_at: float
     resources: str
     status: str
     log_path: str
